Removing old cli code and fixing scalar bug. (#342)
* Removing graphql-transform-cli so as not to confuse and fixing typo in default scalars

* removing old cli and updating tut

+++ packages/graphql-transformer-common/src/definition.ts
@@ -8,7 +8,7 @@ import {
 export const STANDARD_SCALARS = {
     String: 'String',
     Int: 'Int',
-    Float: 'Int',
+    Float: 'Float',
     Boolean: 'Boolean',
     ID: 'ID'
 };
